Tests/LibWeb/Text/expected/css/transition-longhand-var-serialization.txt

.a { transition-property: var(--a); transition-duration: var(--b); transition-timing-function: var(--c); transition-delay: var(--d); transition-behavior: var(--e); }
